// src/core/models/systemsettings.ts
export interface SystemSettingsModel {
  adminGroupId?: string;
  blockMessage?: string;
  erisemiyorumMesaj?: string;
  updatedInacessibleMessage?: string;
  updatedUnblockedMessage?: string;
  updatedStillBlockedMessage?: string;
  hasRedirectionMessage?: string;
  updatedHasRedirectionMessage?: string;
  certificateErrorMessage?: string;
  updatedCertificateErrorMessage?: string;
  telegramBotUserName?: string;
}